major update with bridge options set to on.
Wont display bridges on torrclnd, just on off torrc default if they are already configured, in the sense it wond show on the whiptail (line 419), but will ask if the user want to change.
Either way, this isnt supposed to happen, as both tor istances should be updated with bridges, else there is no reason for bridges. Also, it doesnt matter if it is the same bridges or not, it will work.