Table of options on how to have blitz
I just think this would encourage more people to build the sd card, being clear of pros and cons. I think the way it is structured, it is not clear there is this other option to build, most people don't search enough about this, and it was at the very bottom of the page.