correct bug on debug
What happens is that if a service is `off`, it will still show the logs. The if statement needs to be base if it is `on` (it was, I introduced this bug on a PR a time ago, but just realized it now with v1.7.0 (So sorry, I was an enthusiastic at that time). This goes back to the services part of  v1.6 plus LIT service. Also, added torsocks to termbin).

To understand: Compare [what is happening](http://kq3nkqhpcb3qwxy7nc45hvimruchynvvhmyce4g2ohvcq5pqzl3ih3id.onion/nyxnor/debug.txt) with [my mods now](http://kq3nkqhpcb3qwxy7nc45hvimruchynvvhmyce4g2ohvcq5pqzl3ih3id.onion/nyxnor/debugmods.txt). It is my onion service for testing, the keys is 5CGIZWMBEPY342OJ5BEL4KKPZELCSR66GV57LVK3HOZFB32FSFJQ. This key will change soon.