add client authorization to onion services
https://community.torproject.org/onion-services/advanced/client-auth/
https://github.com/rootzoll/raspiblitz/issues/2306#issuecomment-843562208

New keys overwrite previous one if they are created with the script.
auth on $service - generate 1 key. (I see no reason to loop creating keys as Blitz is supposed to be manage by only one person)
auth off $service - deletes all the keys

I did not repeat the key for services with Tor daemon in the code example at the end because it was not fitting the screen.