Recognizing Armbian system and  fixing "Bitcoin Install Fails" on Armbian (#4157)
* Update build_sdcard.sh

OrangePi5 was not recognized in line ~255; so recognize it by same 'Debian' method, replacing by 'Armbian'

* Update bitcoin.install.sh

Bitcoin Install Fails during build_sdcard.sh due to PGP key download; see https://github.com/raspiblitz/raspiblitz/issues/3753. Comment it by now

* Update bitcoin.install.sh

changed to secure hkps

* Update home.admin/config.scripts/bitcoin.install.sh

These comments are not needed here. The sole purpose was to describe the move to a PR.

Co-authored-by: openoms <43343391+openoms@users.noreply.github.com>

* Update build_sdcard.sh

Modified the request by OpenOMS where the comments are unnecessary and kept Debian as the first option and moved Armbian down the list as an elif.

---------

Co-authored-by: openoms <43343391+openoms@users.noreply.github.com>