use BindsTo in service
PartOf=: Propagates stop and restart commands from the "master" service to the dependent service, but does not automatically start the dependent service when the master service starts.
BindsTo=: Propagates stop and restart commands like PartOf=, but also ensures that when the "master" service starts or restarts, the dependent service also starts automatically.